rename and tidy up
