+++ libs/@hashintel/type-editor/src/entity-type-editor/shared/insert-property-field/type-selector.tsx
